Update 1.0
So... how did we find ourselves at Version 1 before all the features were completed? That is a very good question. I got bored and decided to implement the KeyListener, so technically we have completed the requirements for Version 2 but not  Version 1. So once I finish Version 1, I will launch Version 2. It's confusing, I know, my brain is fried. Along with the KeyListener, numberous bugs were fixed along with the beginnings of status ailments. Happy trails!